api/requirements.txt
Flask==0.10.1
ddtrace==0.12.1
blinker==1.4
datadog
json-logging
gunicorn
psycopg2-binary

